requirements.txt
blinker
click
colorama
contourpy
cycler
et_xmlfile
Flask
Flask-SQLAlchemy
fonttools
greenlet
itsdangerous
Jinja2
kiwisolver
MarkupSafe
matplotlib
numpy
openpyxl
packaging
pandas
pillow
pyparsing
python-dateutil
python-dotenv
pytz
scipy
six
SQLAlchemy
typing_extensions
tzdata
Werkzeug
xlsxwriter
sanic